+++ Plugins/Util/Util.cpp
@@ -2,6 +2,9 @@
 
 #include "API/Constants.hpp"
 #include "API/Globals.hpp"
+#include "API/C2DA.hpp"
+#include "API/CNWRules.hpp"
+#include "API/CTwoDimArrays.hpp"
 #include "API/CResRef.hpp"
 #include "API/CExoResMan.hpp"
 #include "API/CExoString.hpp"
@@ -64,6 +67,7 @@ Util::Util(const Plugin::CreateParams& params)
     REGISTER(GetMinutesPerHour);
     REGISTER(SetMinutesPerHour);
     REGISTER(EncodeStringForURL);
+    REGISTER(Get2DARowCount);
 
 #undef REGISTER
 
@@ -255,6 +259,15 @@ ArgumentStack Util::EncodeStringForURL(ArgumentStack&& args)
     // **
 
     Services::Events::InsertArgument(stack, result);
+    return stack;
+}
+
+ArgumentStack Util::Get2DARowCount(ArgumentStack&& args)
+{
+    ArgumentStack stack;
+    const auto twodaRef = Services::Events::ExtractArgument<std::string>(args);
+    auto twoda = Globals::Rules()->m_p2DArrays->GetCached2DA(twodaRef.c_str(), true);
+    Services::Events::InsertArgument(stack, twoda ? twoda->m_nNumRows : 0);
     return stack;
 }